LayoutTests/platform/ios-sim/media/video-set-rate-from-pause-expected.txt

Test that setting a non-zero rate causes an async timeupdate event.

EVENT(canplaythrough)
EVENT(canplaythrough)
EVENT(play)
EVENT(play)
EVENT(timeupdate)
END OF TEST
